# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OlaClick Restaurant Orders API",
        "description": "API RESTful para la gestión de órdenes de un restaurante, implementada en Laravel 10 siguiendo Arquitectura Hexagonal (Ports & Adapters). Utiliza PostgreSQL como base de datos, Redis para caché, y está completamente desacoplada de Laravel/Eloquent mediante Domain-Driven Design (DDD).",
        "contact": {
            "name": "Jessika Carvajal",
            "email": "jessikacarvajal29@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local de Desarrollo"
        }
    ],
    "paths": {
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener órdenes activas",
                "description": "Retorna todas las órdenes activas (status != 'delivered'). Utiliza Redis para cachear el resultado con TTL de 30 segundos.",
                "operationId": "getActiveOrders",
                "responses": {
                    "200": {
                        "description": "Lista de órdenes activas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Órdenes activas obtenidas exitosamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Crear nueva orden",
                "description": "Crea una nueva orden con estado inicial 'initiated'. La orden debe incluir el nombre del cliente y al menos un item.",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Orden creada exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener detalle de una orden",
                "description": "Muestra el detalle completo de una orden específica. **Nota: Este endpoint aún no está implementado con la nueva arquitectura hexagonal.**",
                "operationId": "getOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID único de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la orden obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Orden obtenida exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Orden no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Método no implementado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Método no implementado con la nueva arquitectura"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/advance": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Avanzar estado de orden",
                "description": "Avanza el estado de una orden siguiendo el flujo: initiated → sent → delivered. Cuando una orden llega al estado 'delivered', se elimina automáticamente de la base de datos y del caché.",
                "operationId": "advanceOrderStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID único de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado de la orden actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Estado de la orden actualizado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Orden no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Transición de estado inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No se puede avanzar el estado de esta orden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Order": {
                "title": "Order",
                "description": "Orden del restaurante",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "client_name": {
                        "type": "string",
                        "example": "Jessika Carvajal"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "initiated",
                            "sent",
                            "delivered"
                        ],
                        "example": "initiated"
                    },
                    "total_amount": {
                        "type": "string",
                        "example": "85.00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-17T00:14:12.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-17T00:14:12.000000Z"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "title": "OrderItem",
                "description": "Item de una orden",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "type": "string",
                        "example": "Lomo saltado"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 1
                    },
                    "unit_price": {
                        "type": "string",
                        "example": "60.00"
                    }
                },
                "type": "object"
            },
            "CreateOrderRequest": {
                "title": "CreateOrderRequest",
                "description": "Datos para crear una nueva orden",
                "required": [
                    "client_name",
                    "items"
                ],
                "properties": {
                    "client_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Jessika Carvajal"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CreateOrderItemRequest"
                        },
                        "minItems": 1
                    }
                },
                "type": "object"
            },
            "CreateOrderItemRequest": {
                "title": "CreateOrderItemRequest",
                "description": "Datos para crear un item de orden",
                "required": [
                    "description",
                    "quantity",
                    "unit_price"
                ],
                "properties": {
                    "description": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Lomo saltado"
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0.01,
                        "example": 60
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "ErrorResponse",
                "description": "Respuesta de error estándar",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error al procesar la solicitud"
                    },
                    "error": {
                        "type": "string",
                        "example": "Descripción detallada del error"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "title": "ValidationErrorResponse",
                "description": "Respuesta de error de validación",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "client_name": [
                                "El nombre del cliente es obligatorio."
                            ],
                            "items.0.quantity": [
                                "La cantidad debe ser al menos 1."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Operaciones para la gestión de órdenes del restaurante"
        }
    ]
}